# docs/swagger.yaml
basePath: /api
definitions:
  models.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  models.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ApiError'
      success:
        type: boolean
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.Product:
    properties:
      categories:
        description: Joined category data
        items:
          $ref: '#/definitions/models.Category'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
      updated_at:
        type: string
    required:
    - name
    - price
    - stock
    type: object
  models.ProductCreateRequest:
    properties:
      category_ids:
        description: At least one category
        items:
          type: integer
        minItems: 1
        type: array
      description:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_ids
    - name
    - price
    - stock
    type: object
  models.ProductHistory:
    properties:
      changed_at:
        type: string
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      stock:
        type: integer
    type: object
  models.ProductHistoryResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/models.ProductHistory'
        type: array
      total:
        type: integer
    type: object
  models.ProductListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  models.ProductUpdateRequest:
    properties:
      category_ids:
        items:
          type: integer
        minItems: 1
        type: array
      description:
        type: string
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: bruno.malagoli@example.com
    name: Bsmart API Support
  description: RESTful API con soporte de WebSocket en tiempo real para gestión de
    productos y categorías. Sistema completo con autenticación JWT, control de acceso
    basado en roles, búsqueda full-text y seguimiento automático de historial de productos.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bsmart Backend API
  version: "1.0"
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de productos con opciones de filtrado,
        ordenamiento y búsqueda full-text
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 10
        description: Productos por página (máximo 100)
        in: query
        name: limit
        type: integer
      - default: created_at
        description: Campo para ordenar
        enum:
        - name
        - price
        - stock
        - created_at
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Dirección de ordenamiento
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Término de búsqueda full-text en nombres de productos
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de productos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductListResponse'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Listar productos con paginación
      tags:
      - productos
    post:
      consumes:
      - application/json
      description: Crea un nuevo producto en el sistema. Requiere rol de administrador.
        Emite evento WebSocket 'product:created'.
      parameters:
      - description: Datos del producto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProductCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Producto creado exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Crear nuevo producto
      tags:
      - productos
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un producto del sistema. Requiere rol de administrador.
        Emite evento WebSocket 'product:deleted'.
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Producto eliminado exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: ID inválido
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Producto no encontrado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Eliminar producto
      tags:
      - productos
    get:
      consumes:
      - application/json
      description: Obtiene la información completa de un producto específico incluyendo
        sus categorías
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalle del producto
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: ID inválido
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Producto no encontrado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener detalle de producto
      tags:
      - productos
    put:
      consumes:
      - application/json
      description: Actualiza un producto existente. Requiere rol de administrador.
        Los campos no enviados no se modifican. Emite evento WebSocket 'product:updated'.
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar (campos opcionales)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ProductUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Producto actualizado exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Product'
              type: object
        "400":
          description: ID inválido o datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Producto no encontrado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Actualizar producto
      tags:
      - productos
  /products/{id}/history:
    get:
      consumes:
      - application/json
      description: Obtiene el historial de cambios de precio y stock de un producto.
        Opcionalmente filtra por rango de fechas (formato RFC3339).
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: integer
      - description: 'Fecha de inicio (RFC3339: 2024-01-01T00:00:00Z)'
        in: query
        name: start
        type: string
      - description: 'Fecha de fin (RFC3339: 2024-12-31T23:59:59Z)'
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Historial del producto
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductHistoryResponse'
              type: object
        "400":
          description: ID inválido o formato de fecha incorrecto
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener historial de producto
      tags:
      - productos
...
securityDefinitions:
  BearerAuth:
    description: 'Autenticación JWT. Formato: "Bearer {token}". Obtén un token desde
      /api/auth/login o /api/auth/register'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"